+++ CHANGELOG.md
@@ -4,14 +4,16 @@
 ### Fixed
 - `pyinseq` alone brings up the help documentation
 - Small fix to the `three_primeness` calculation.
-- Added progress bar for `demultiplex` function
-- Added a fastq file generator with mmap
+
 
 ### Changed
 - Only Python 3.6+ supported
+
 ### Added
 - `pyinseq genomeprep` command
 - Add T50 calculation
+- Added progress bar for `demultiplex` function
+- Added a fastq file generator with mmap
 
 ## [0.2.0] - 2017-07-16
 ### Added